Refactor shell into multiple modules
