+++ src/builtins.h
@@ -1,0 +1,6 @@
+#ifndef BUILTINS_H
+#define BUILTINS_H
+
+int run_builtin(char **args);
+
+#endif /* BUILTINS_H */
