+++ src/main.c
@@ -6,190 +6,9 @@
 #include <sys/wait.h>
 #include <signal.h>
 
-#define MAX_TOKENS 64
-#define MAX_LINE 1024
-
-typedef struct Job {
-    int id;
-    pid_t pid;
-    char cmd[MAX_LINE];
-    struct Job *next;
-} Job;
-
-static Job *jobs = NULL;
-static int next_job_id = 1;
-
-static void add_job(pid_t pid, const char *cmd) {
-    Job *job = malloc(sizeof(Job));
-    if (!job) return;
-    job->id = next_job_id++;
-    job->pid = pid;
-    strncpy(job->cmd, cmd, MAX_LINE - 1);
-    job->cmd[MAX_LINE - 1] = '\0';
-    job->next = jobs;
-    jobs = job;
-}
-
-static void remove_job(pid_t pid) {
-    Job **curr = &jobs;
-    while (*curr) {
-        if ((*curr)->pid == pid) {
-            Job *tmp = *curr;
-            *curr = (*curr)->next;
-            free(tmp);
-            return;
-        }
-        curr = &((*curr)->next);
-    }
-}
-
-static void check_jobs(void) {
-    int status;
-    pid_t pid;
-    while ((pid = waitpid(-1, &status, WNOHANG)) > 0) {
-        printf("[vush] job %d finished\n", pid);
-        remove_job(pid);
-    }
-}
-
-static void print_jobs(void) {
-    Job *j = jobs;
-    while (j) {
-        printf("[%d] %d %s\n", j->id, j->pid, j->cmd);
-        j = j->next;
-    }
-}
-
-static char *expand_var(const char *token) {
-    if (token[0] == '~') {
-        const char *home = getenv("HOME");
-        if (!home) home = "";
-        size_t home_len = strlen(home);
-        size_t token_len = strlen(token);
-        char *tmp = malloc(home_len + token_len);
-        if (!tmp) return NULL;
-        strcpy(tmp, home);
-        strcat(tmp, token + 1);
-        char *ret = strdup(tmp);
-        free(tmp);
-        return ret;
-    }
-    if (token[0] != '$') return strdup(token);
-    const char *val = getenv(token + 1);
-    return strdup(val ? val : "");
-}
-
-static int parse_line(char *line, char **args, int *background) {
-    int argc = 0;
-    *background = 0;
-    char *p = line;
-    while (*p && argc < MAX_TOKENS - 1) {
-        while (*p == ' ' || *p == '\t') p++;
-        if (*p == '\0') break;
-
-        char buf[MAX_LINE];
-        int len = 0;
-        int do_expand = 1;
-
-        if (*p == '\'') {
-            /* single quoted token - copy verbatim */
-            do_expand = 0;
-            p++; /* skip opening quote */
-            while (*p && *p != '\'' && len < MAX_LINE - 1) {
-                buf[len++] = *p++;
-            }
-            if (*p == '\'') p++; /* skip closing quote */
-        } else {
-            int in_double = 0;
-            if (*p == '"') {
-                in_double = 1;
-                p++; /* skip opening quote */
-            }
-            int first = 1;
-            while (*p && (in_double || (*p != ' ' && *p != '\t'))) {
-                if (*p == '\\') {
-                    p++;
-                    if (*p) {
-                        if (len < MAX_LINE - 1) buf[len++] = *p;
-                        if (first) do_expand = 0;
-                        if (*p) p++;
-                    }
-                    first = 0;
-                    continue;
-                }
-                if (in_double && *p == '"') {
-                    p++; /* end quote */
-                    break;
-                }
-                if (len < MAX_LINE - 1) buf[len++] = *p;
-                p++;
-                first = 0;
-            }
-        }
-
-        buf[len] = '\0';
-        args[argc++] = do_expand ? expand_var(buf) : strdup(buf);
-    }
-
-    if (argc > 0 && strcmp(args[argc-1], "&") == 0) {
-        *background = 1;
-        free(args[argc-1]);
-        argc--;
-    }
-
-    args[argc] = NULL;
-    return argc;
-}
-
-static int builtin_cd(char **args) {
-    const char *dir = args[1] ? args[1] : getenv("HOME");
-    if (chdir(dir) != 0) {
-        perror("cd");
-    }
-    return 1;
-}
-
-static int builtin_exit(char **args) {
-    (void)args;
-    exit(0);
-}
-
-static int builtin_pwd(char **args) {
-    (void)args;
-    char cwd[1024];
-    if (getcwd(cwd, sizeof(cwd))) {
-        printf("%s\n", cwd);
-    } else {
-        perror("pwd");
-    }
-    return 1;
-}
-
-static int builtin_jobs(char **args) {
-    (void)args;
-    print_jobs();
-    return 1;
-}
-
-static struct builtin {
-    const char *name;
-    int (*func)(char **);
-} builtins[] = {
-    {"cd", builtin_cd},
-    {"exit", builtin_exit},
-    {"pwd", builtin_pwd},
-    {"jobs", builtin_jobs},
-    {NULL, NULL}
-};
-
-static int run_builtin(char **args) {
-    for (int i = 0; builtins[i].name; i++) {
-        if (strcmp(args[0], builtins[i].name) == 0) {
-            return builtins[i].func(args);
-        }
-    }
-    return 0;
-}
+#include "parser.h"
+#include "jobs.h"
+#include "builtins.h"
 
 int main(void) {
     char line[MAX_LINE];
